+++ Gdax/Authentication/IAuthenticator.cs
@@ -1,0 +1,7 @@
+﻿namespace Gdax.Authentication
+{
+	public interface IAuthenticator
+	{
+		AuthenticationToken GetAuthenticationToken(GdaxRequest request);
+	}
+}