+++ Gdax/Class1.cs
@@ -1,6 +1,0 @@
-﻿namespace Gdax
-{
-	public class Class1
-	{
-	}
-}