+++ Gdax/DateTimeExtensions.cs
@@ -1,0 +1,12 @@
+﻿using System;
+
+namespace Gdax
+{
+	internal static class DateTimeExtensions
+	{
+		public static Double ToUnixTimestamp(this DateTime dateTime)
+		{
+			return (dateTime - new DateTime(1970, 1, 1, 0, 0, 0, DateTimeKind.Utc)).TotalSeconds;
+		}
+	}
+}