+++ Gdax/Models/Time.cs
@@ -1,0 +1,10 @@
+﻿using System;
+
+namespace Gdax.Models
+{
+	public class Time
+	{
+		public DateTime Iso { get; set; }
+		public Decimal Epoch { get; set; }
+	}
+}