+++ CHANGELOG.md
@@ -5,5 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 ## [Unreleased]
 ### Added
+- `GdaxAuthenticator` for authenticating to GDAX via API keys
+- The following Api's have been added:
+  - `GdaxClient.GetServerTimeAsync`
 
 [Unreleased]: https://github.com/ChristopherHaws/gdax-dotnet/compare/v1.0.0...HEAD